Move engine into devserver package
The engine is tied directly to the development server;  it is not
capable of being used outside of a development environment.  To this
effect, we move the engnine into the devserver package.

Note that this is still an exported member (for the time being).
