+++ pkg/devserver/devserver.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/inngest/inngestctl/pkg/api"
-	"github.com/inngest/inngestctl/pkg/engine"
 	"github.com/rs/zerolog"
 )
 
@@ -16,7 +15,7 @@ type Options struct {
 
 type DevServer struct {
 	API    api.API
-	Engine *engine.Engine
+	Engine *Engine
 
 	dir string
 	log *zerolog.Logger
@@ -24,9 +23,7 @@ type DevServer struct {
 
 // Create and start a new dev server (API, Exectutor, State, Logger, etc.)
 func NewDevServer(o Options) (DevServer, error) {
-	eng, err := engine.New(engine.Options{
-		Logger: o.Log,
-	})
+	eng, err := NewEngine(o.Log)
 	if err != nil {
 		return DevServer{}, err
 	}
