+++ pkg/devserver/logging_queue.go
@@ -1,4 +1,4 @@
-package engine
+package devserver
 
 import (
 	"context"
